requirements.txt
flask
flask-cors
tensorflow
pillow
numpy
scikit-learn
requests
psutil